# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OTel Data API",
    "description": "REST API for OwnTracks/Garmin GPS data with PostGIS spatial queries",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/gps/unified": {
      "get": {
        "tags": [
          "Unified GPS"
        ],
        "summary": "List Unified Gps",
        "description": "Query the unified_gps_points view combining OwnTracks + Garmin data.\n\nMerges OwnTracks location data and Garmin track points into a single\nchronological stream. Filter by data source or date range.",
        "operationId": "list_unified_gps_api_v1_gps_unified_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source: owntracks or garmin",
              "examples": [
                "owntracks"
              ],
              "title": "Source"
            },
            "description": "Filter by source: owntracks or garmin"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter from date (YYYY-MM-DD)",
              "examples": [
                "2026-02-01"
              ],
              "title": "Date From"
            },
            "description": "Filter from date (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to date (YYYY-MM-DD)",
              "examples": [
                "2026-02-12"
              ],
              "title": "Date To"
            },
            "description": "Filter to date (YYYY-MM-DD)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UnifiedGpsPoint_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gps/daily-summary": {
      "get": {
        "tags": [
          "Unified GPS"
        ],
        "summary": "Daily Summary",
        "description": "Query the daily_activity_summary view for aggregated daily stats.\n\nReturns per-day aggregates including OwnTracks point counts, battery stats,\nand Garmin activity metrics (distance, duration, heart rate, calories).",
        "operationId": "daily_summary_api_v1_gps_daily_summary_get",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter from date (YYYY-MM-DD)",
              "examples": [
                "2026-02-01"
              ],
              "title": "Date From"
            },
            "description": "Filter from date (YYYY-MM-DD)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to date (YYYY-MM-DD)",
              "examples": [
                "2026-02-12"
              ],
              "title": "Date To"
            },
            "description": "Filter to date (YYYY-MM-DD)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyActivitySummary"
                  },
                  "title": "Response Daily Summary Api V1 Gps Daily Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DailyActivitySummary": {
        "properties": {
          "activity_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Date"
          },
          "owntracks_device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owntracks Device"
          },
          "owntracks_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owntracks Points"
          },
          "min_battery": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Battery"
          },
          "max_battery": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Battery"
          },
          "avg_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Accuracy"
          },
          "garmin_sport": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Garmin Sport"
          },
          "garmin_activities": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Garmin Activities"
          },
          "total_distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Distance Km"
          },
          "total_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Seconds"
          },
          "avg_heart_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Heart Rate"
          },
          "total_calories": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Calories"
          }
        },
        "type": "object",
        "title": "DailyActivitySummary",
        "examples": [
          {
            "activity_date": "2026-02-11",
            "avg_accuracy": 4.89,
            "avg_heart_rate": 142.0,
            "garmin_activities": 1,
            "garmin_sport": "cycling",
            "max_battery": 100,
            "min_battery": 64,
            "owntracks_device": "iphone_stuart",
            "owntracks_points": 1429,
            "total_calories": 1689,
            "total_distance_km": 50.6,
            "total_duration_seconds": 6932
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_UnifiedGpsPoint_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UnifiedGpsPoint"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[UnifiedGpsPoint]"
      },
...
      "UnifiedGpsPoint": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "battery": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "heart_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "source",
          "identifier",
          "latitude",
          "longitude",
          "timestamp"
        ],
        "title": "UnifiedGpsPoint",
        "examples": [
          {
            "accuracy": 7,
            "battery": 100,
            "created_at": "2026-02-12T08:11:55Z",
            "identifier": "iphone_stuart",
            "latitude": 40.736238,
            "longitude": -74.039405,
            "source": "owntracks",
            "timestamp": "2026-02-12T08:11:55Z"
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}